Add fourth task and README.md

+++ src/main/java/trainingday1/thirdtask/AreaOfSquareInscribedInCircle.java
@@ -2,7 +2,7 @@ package trainingday1.thirdtask;
 
 public class AreaOfSquareInscribedInCircle {
 
-    public int inscribedSquare(int outerSquare){
+    public double inscribedSquare(double outerSquare){
         return outerSquare/2;
     }
     public static void main (String[]args){
